requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
contextlib2==21.6.0
fandom-py==0.2.1
idna==3.3
oauthlib==3.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
schema==0.7.4
soupsieve==2.2.1
tweepy==4.2.0
urllib3==1.26.7
